feat(): clean up update/retrieve schema

+++ packages/api/schema/hubify/users.ts
@@ -9,7 +9,7 @@ export const columns = defineTableColumns({
     notNull: true
   },
   password: {
-    type: 'int8',
+    type: 'int4',
     notNull: false
   },
   createdAt: {
